change class game
