+++ src/com/newgame/xo/model/Game.java
@@ -10,7 +10,7 @@ public class Game {
 
     private final String name;
 
-    Game(final GameBuilder builder) {
+    Game(final Builder builder) {
         this.player1 = builder.getPlayer1();
         this.player2 = builder.getPlayer2();
         this.field = builder.getField();
@@ -31,5 +31,60 @@ public class Game {
 
     public String getName() {
         return name;
+    }
+
+    public static class Builder {
+        public Player player1;
+
+        public Player player2;
+
+        public Field field;
+
+        public String name;
+
+        public Player getPlayer1() {
+            return this.player1;
+        }
+
+        public Player getPlayer2() {
+            return this.player2;
+        }
+
+        public Field getField() {
+            return this.field;
+        }
+
+        public String getName() {
+            return this.name;
+        }
+
+        public Builder player1(final Player player1) {
+            this.player1 = player1;
+            return this;
+        }
+
+        public Builder player2(final Player player2) {
+            this.player2 = player2;
+            return this;
+        }
+
+        public Builder field(final Field field) {
+            this.field = field;
+            return this;
+        }
+
+        public Builder name(final String name) {
+            this.name = name;
+            return this;
+        }
+
+        public Game build() {
+            return new Game(this);
+        }
+
+
+
+
+
     }
 }
