+++ src/com/newgame/xo/view/AICoordinateGetter.java
@@ -6,7 +6,7 @@ import com.newgame.xo.model.Point;
 
 import java.util.Random;
 
-public class AICoordinateGetter implements ICoordinateGetter {
+public class AICoordinateGetter extends ICoordinateGetter {
   // полу рандомный ход с центра
 
     public Point getMoveCoordinate(final Field field) {
