+++ src/com/newgame/xo/view/ICoordinateGetter.java
@@ -3,9 +3,9 @@ package com.newgame.xo.view;
 import com.newgame.xo.model.Field;
 import com.newgame.xo.model.Point;
 
-public interface ICoordinateGetter {
+public abstract class ICoordinateGetter {
 
-    Point getMoveCoordinate(final Field field);
+    public abstract Point getMoveCoordinate(final Field field);
 
 }
